+++ pe/_meta.py
@@ -2,6 +2,6 @@
 Meta-information about pe.
 """
 
-__version__ = '0.3.0'
+__version__ = '0.3.1'
 
 # Note: most of this file's contents has moved to setup.cfg